+++ cpp/locality/NeighborComputeFunctional.cc
@@ -32,9 +32,9 @@ NeighborList makeDefaultNlist(const NeighborQuery* nq, const NeighborList* nlist
     locality::NeighborList new_nlist = NeighborList(*nlist);
     new_nlist.validate(num_query_points, nq->getNPoints());
     if (requires_delete)
-        {
+    {
         delete nlist;
-        }
+    }
     return new_nlist;
 }