+++ tests/test_density_GaussianDensity.py
@@ -74,34 +74,62 @@ class TestGaussianDensity(unittest.TestCase):
             gd.compute((test_box, test_points))
 
     def test_sum_2d(self):
-        # Ensure that the Gaussian sums to 1
+        # Ensure that each point's Gaussian sums to 1
         width = 100
         r_max = 49
         sigma = 10
-        num_points = 1
         box_size = width
-        box, points = freud.data.make_random_system(
-            box_size, num_points, is2D=True)
-
         gd = freud.density.GaussianDensity(width, r_max, sigma)
-        gd.compute(system=(box, points))
-        # This has discretization error as well as single-precision error
-        assert np.isclose(np.sum(gd.density), 1, atol=1e-4)
+        for num_points in [1, 10, 100]:
+            box, points = freud.data.make_random_system(
+                box_size, num_points, is2D=True)
+            gd.compute(system=(box, points))
+            # This has discretization error as well as single-precision error
+            assert np.isclose(np.sum(gd.density), num_points, rtol=1e-4)
 
     def test_sum_3d(self):
-        # Ensure that the Gaussian sums to 1
+        # Ensure that each point's Gaussian sums to 1
         width = 100
         r_max = 49
         sigma = 10
-        num_points = 1
         box_size = width
-        box, points = freud.data.make_random_system(
-            box_size, num_points, is2D=False)
-
         gd = freud.density.GaussianDensity(width, r_max, sigma)
-        gd.compute(system=(box, points))
-        # This has discretization error as well as single-precision error
-        assert np.isclose(np.sum(gd.density), 1, atol=1e-4)
+        for num_points in [1, 10, 100]:
+            box, points = freud.data.make_random_system(
+                box_size, num_points, is2D=False)
+            gd.compute(system=(box, points))
+            # This has discretization error as well as single-precision error
+            assert np.isclose(np.sum(gd.density), num_points, rtol=1e-4)
+
+    def test_sum_values_2d(self):
+        # Ensure that the Gaussian convolution sums to the sum of the values
+        width = 100
+        r_max = 49
+        sigma = 10
+        box_size = width
+        gd = freud.density.GaussianDensity(width, r_max, sigma)
+        for num_points in [1, 10, 100]:
+            system = freud.data.make_random_system(
+                box_size, num_points, is2D=True)
+            values = np.random.rand(num_points)
+            gd.compute(system, values)
+            # This has discretization error as well as single-precision error
+            assert np.isclose(np.sum(gd.density), np.sum(values), rtol=1e-4)
+
+    def test_sum_values_3d(self):
+        # Ensure that the Gaussian convolution sums to the sum of the values
+        width = 100
+        r_max = 49
+        sigma = 10
+        box_size = width
+        gd = freud.density.GaussianDensity(width, r_max, sigma)
+        for num_points in [1, 10, 100]:
+            system = freud.data.make_random_system(
+                box_size, num_points, is2D=False)
+            values = np.random.rand(num_points)
+            gd.compute(system, values)
+            # This has discretization error as well as single-precision error
+            assert np.isclose(np.sum(gd.density), np.sum(values), rtol=1e-4)
 
     def test_repr(self):
         gd = freud.density.GaussianDensity(100, 10.0, 0.1)